Differential Reward system added
Ball ids are set to respond appropriately to their respective players this works for all wl conditions and the balls are now colour coord with p's paddles as well.

The differential rewards are now working but I imagine the data it outputs is still not 100% need to discuss with murray still how to get data files etc and I will also need to create a new variable that holds hits specifically for the rewards. At least from participant end you can see how this will work.